requirements.txt
anyio==3.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.1
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
peewee==3.16.2
pluggy==1.0.0
protobuf==3.20.3
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
